Modified the types of matrix or vector sizes and indices, so as to be consistent with the latest versions of Seldon in which 'int' has been replaced with 'size_t'.

+++ method/BLUE.cxx
@@ -377,8 +377,8 @@ namespace Verdandi
     {
         typedef typename State::value_type T;
 
-        int Ny = y.GetLength();
-        int Nx = x.GetLength();
+        size_t Ny = y.GetLength();
+        size_t Nx = x.GetLength();
 
         if (B.GetM() != Nx || B.GetN() != Nx)
             throw ErrorArgument("ComputeBLUE_matrix",
